backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "weaviate-client",
    "pdf2image",
    "pillow",
    "python-dotenv",
    "google-genai",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "supabase",
    "audio-recorder-streamlit>=0.0.10",
    "railtracks[cli]>=1.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "hypothesis",
    "reportlab",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.9",
    "reportlab>=4.4.10",
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests requiring live external services (deselect with '-m not integration')",
]
